Updated create ad behaviour on Operator Dashboard
When reloading the create form because of an error the controller now checks to see if an existing Advertiser or Ad Campaign was selected. If so it retrieves the appropriate model objects and adds them to the form. The result of this means that the Logo for an existing Advertiser will be properly displayed now upon reload.